Add files via upload
As a beginner hacker at the beginning of this Hackathon, I was really scared that I would not be able to fit in. However, upon learning new skills, I learned to do little projects like the one I am submitting here. Although it is not a full project, I have put in a lot of effort, and am proud of it.